Only clean dir if it exists in FileSourceModuleTests
If other initialization fails, we get an IllegalArgumentException
from the AfterClass method.

+++ spring-xd-dirt/src/test/java/org/springframework/xd/dirt/stream/FileSourceModuleTests.java
@@ -58,7 +58,9 @@ public class FileSourceModuleTests extends StreamTestSupport {
 
 	@Before
 	public void setUp() throws IOException {
-		FileUtils.cleanDirectory(sourceDir);
+		if (sourceDir.exists()) {
+			FileUtils.cleanDirectory(sourceDir);
+		}
 	}
 
 	@Test
@@ -130,7 +132,9 @@ public class FileSourceModuleTests extends StreamTestSupport {
 
 	@AfterClass
 	public static void deleteTempDir() throws IOException {
-		FileUtils.cleanDirectory(sourceDir);
-		FileUtils.deleteDirectory(sourceDir);
+		if (sourceDir.exists()) {
+			FileUtils.cleanDirectory(sourceDir);
+			FileUtils.deleteDirectory(sourceDir);
+		}
 	}
 }
